move resolver
