+++ flytekit/core/class_based_resolver.py
@@ -1,7 +1,8 @@
 from typing import List
 
+from flytekit.core.base_task import TaskResolverMixin
 from flytekit.core.context_manager import SerializationSettings
-from flytekit.core.python_auto_container import PythonAutoContainerTask, TaskResolverMixin
+from flytekit.core.python_auto_container import PythonAutoContainerTask
 from flytekit.core.tracker import TrackedInstance
