+++ flytekit/extend/__init__.py
@@ -32,11 +32,10 @@ This package contains things that are useful when extending Flytekit.
 from flytekit.common.translator import get_serializable
 from flytekit.core import context_manager
 from flytekit.core.base_sql_task import SQLTask
-from flytekit.core.base_task import IgnoreOutputs, PythonTask
+from flytekit.core.base_task import IgnoreOutputs, PythonTask, TaskResolverMixin
 from flytekit.core.class_based_resolver import ClassStorageTaskResolver
 from flytekit.core.context_manager import ExecutionState, Image, ImageConfig, SerializationSettings
 from flytekit.core.interface import Interface
 from flytekit.core.promise import Promise
-from flytekit.core.python_auto_container import TaskResolverMixin
 from flytekit.core.task import TaskPlugins
 from flytekit.core.type_engine import DictTransformer, T, TypeEngine, TypeTransformer
